brightness: Allow the user to specify the brightness percentage.
The granularity of the percentage seems to be machine dependent.

On one notebook that I have, it seems to be in steps of 12.5%, but as it
seems to need to be an integer, it only works when using the ceiling of the
percentage, that is, to get a brightness equivalent to 62.5%, I have to
specify 63 as the parameter.