variance: Add script to calculate the (sample) variance of a list of reals.
The algorithm used here is Welford's recurrence formula. As a side-product,
we also calculate the *sample* mean of the list of integers given as input.

Note that we use only $O(1)$ operations per each iteration for a total of
$n$ iterations, which means that we compute the variance in $\Theta(n)$
time, where $n$ is the number of input numbers (the input list).

Another nice property of this algorithm is that it is an on-line algorithm,
meaning that it does not need to have all the data present before it starts
the computation. The current implementation is not on-line, but it can
simply be modified to be one with the use of generators.