optimize_jpegs: Use a generator to make iteration more pythonic.
This also allows us to remove one level of indentation, which makes the code
more readable.

The best alternative would be for the pikepdf library to expose iteration of
the images out-of-the-box. This would execute native code instead of
executing interpreted (Python) code.

Another (better?) alternative would be to write a derived/adapter class
based on pikepdf that has this generator/iterator of image objects as a
method, but it is probably not worth the effort (this needs to be
evaluated).