using_pikepdf: Report number of saved bytes.
This may be a negative value if the original file was better compressed than
pikepdf/qpdf compresses, but further processing with pdfsizeopt in almost
all cases recovers the previous gains (and further improves upon them, since
we have fewer data to store).

The best data compression, after all is the omission of data. ☺