using_pikepdf: Refactor opening file + deleting metadata.
We will actually want to open the file only once to perform multiple
optimizations.

Apparently, pikepdf parses the PDF file lazily, which means that the second
time that it iterates over the file, it does so much more quickly.

We can rewrite (later) the loop to iterate over the objects only once (and
dispatch the appropriate function, depending on the object) if the need
arises.